refactor: 구매가격을 Long형으로 변경

+++ src/main/java/roomescape/dto/business/PaymentCreationContent.java
@@ -6,7 +6,7 @@ import roomescape.dto.request.WaitingCreationRequest;
 public record PaymentCreationContent(
         String orderId,
         String paymentKey,
-        Integer amount
+        Long amount
 ) {
 
     public PaymentCreationContent(ReservationCreationRequest request) {
